+++ lib/express.js
@@ -42,21 +42,23 @@ function createApplication() {
   mixin(app, proto, false);
 
   // expose the prototype that will get set on requests
-  app.request = Object.create(req.req, {
+  app.request = Object.create(req, {
     app: { configurable: true, enumerable: true, writable: true, value: app }
   })
 
   // expose the prototype that will get set on responses
-  app.response = Object.create(res.res, {
+  app.response = Object.create(res, {
     app: { configurable: true, enumerable: true, writable: true, value: app }
   })
 
   if (isHttp2Supported) {
-    app.http2Request = Object.create(req.http2Req, {
+    var http2Req = require('./http2Request');
+    var http2Res = require('./http2Response');
+    app.http2Request = Object.create(http2Req, {
       app: { configurable: true, enumerable: true, writable: true, value: app }
     });
 
-    app.http2Response = Object.create(res.http2Res, {
+    app.http2Response = Object.create(http2Res, {
       app: { configurable: true, enumerable: true, writable: true, value: app }
     });
   }
