+++ lib/utils.js
@@ -288,6 +288,23 @@ exports.setCharset = function setCharset(type, charset) {
 };
 
 /**
+ * Helper function for creating a getter on an object.
+ *
+ * @param {Object} obj
+ * @param {String} name
+ * @param {Function} getter
+ * @private
+ */
+
+exports.defineGetter = function defineGetter(obj, name, getter) {
+  Object.defineProperty(obj, name, {
+    configurable: true,
+    enumerable: true,
+    get: getter
+  });
+}
+
+/**
  * Parse an extended query string with qs.
  *
  * @return {Object}
